Split off yolov3 tasks into their own project (and split off anything that needed)
