+++ core/dsl-test-util/src/main/kotlin/edu/wpi/axon/dsl/MockUtil.kt
@@ -5,6 +5,8 @@ import edu.wpi.axon.dsl.validator.variablename.VariableNameValidator
 import io.mockk.every
 import io.mockk.mockk
 
+// TODO: Find a way to only have one of these files
+
 fun mockVariableNameValidator(vararg nameValidations: Pair<String, Boolean>) =
     mockk<VariableNameValidator> {
         nameValidations.forEach { (name, isValid) ->
