+++ core/tasks-yolov3/tasks-yolov3.gradle.kts
@@ -1,0 +1,9 @@
+description = "Task implementations for yolov3."
+
+dependencies {
+    api(project(":dsl"))
+
+    implementation(project(":util"))
+
+    testImplementation(project(":dsl-test-util"))
+}
